requirements.txt
streamlit
requests
transformers
torch
sentencepiece
python-dotenv
